+++ packages/sealed_languages/lib/src/model/script/writing_system.dart
@@ -1,4 +1,5 @@
 import "../../data/scripts.data.dart";
+import "../../helpers/extensions/iso_standardized_string_extension.dart";
 import "../../helpers/script/script_json.dart";
 import "../../interfaces/iso_standardized.dart";
 import "../../interfaces/json_encodable.dart";
@@ -24,5 +25,5 @@ sealed class WritingSystem implements Named<String> {
   final String name;
 
   @override
-  String toString() => '$WritingSystem(name: "$name")';
+  String toString({bool short = true}) => 'WritingSystem(name: "$name")';
 }