+++ packages/sealed_languages/lib/src/model/translated_name.dart
@@ -57,8 +57,8 @@ class TranslatedName implements Named<String>, JsonEncodable<TranslatedName> {
 
   @override
   String toString({bool short = true}) => short
-      ? '''$TranslatedName(${language.runtimeType}(), name: "$name", fullName: ${fullName == null ? fullName : '"$fullName"'}, countryCode: ${countryCode == null ? countryCode : '"$countryCode"'}, script: ${script == null ? script : '${script?.runtimeType}()'})'''
-      : '''$TranslatedName($language, name: "$name", fullName: ${fullName == null ? fullName : '"$fullName"'}, countryCode: ${countryCode == null ? countryCode : '"$countryCode"'}, script: $script)''';
+      ? name
+      : '''TranslatedName(${language.runtimeType}(), name: "$name", fullName: ${fullName == null ? fullName : '"$fullName"'}, countryCode: ${countryCode == null ? countryCode : '"$countryCode"'}, script: ${script == null ? script : '${script?.runtimeType}()'})''';
 
   @override
   String toJson({JsonCodec codec = const JsonCodec()}) => codec.encode(toMap());